requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
sqlmodel
psycopg2-binary
celery
redis
python-socketio[asyncio_client]
httpx
python-multipart
starlette